Fixed authentication and code refactor
