+++ Backend/Backend.API/Queries/UserQuery.cs
@@ -4,18 +4,20 @@ using HotChocolate.Authorization;
 
 namespace Backend.API.Queries
 {
-    [Authorize]
     [ExtendObjectType("Query")]
     public class UserQuery ([Service] IUserRepository userRepository)
     {
+        [AllowAnonymous]
         public Task<IEnumerable<User>> GetUsersAsync() => userRepository.GetAllAsync();
 
+        [AllowAnonymous]
         public async Task<bool> UserExistByEmail(string email)
         {
             User user = await userRepository.GetByEmailAsync(email);
             return user != null;
         }
 
+        [Authorize]
         public async Task<User> GetUserByEmail(string email)
         {
             var res = await userRepository.GetByEmailAsync(email);
