+++ Backend/Backend.Core/Backend.Core.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MongoDB.Bson" Version="2.23.1" />
+    <PackageReference Include="MongoDB.Bson" Version="2.23.2" />
   </ItemGroup>
 
 </Project>
