+++ Backend/Backend.Infrastructure/Backend.Infrastructure.csproj
@@ -8,7 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Options" Version="8.0.2" />
-    <PackageReference Include="MongoDB.Driver" Version="2.23.1" />
+    <PackageReference Include="MongoDB.Driver" Version="2.23.2" />
   </ItemGroup>
 
   <ItemGroup>
