+++ frontend/app/components/dashboard/transactionCard.tsx
@@ -1,12 +1,7 @@
+"use client";
+
 import { useUserStore } from "@/utils/zustand/userStore";
-import {
-  Card,
-  CardBody,
-  Stat,
-  StatArrow,
-  StatLabel,
-  StatNumber,
-} from "@chakra-ui/react";
+import { Card, CardBody, Stat, StatArrow, StatLabel, StatNumber } from "@chakra-ui/react";
 
 const TransactionCard = ({ isIncome }: { isIncome: boolean }) => {
   const { currency, expenseAmount, incomeAmount } = useUserStore();
