+++ frontend/app/dashboard/page.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { Box, Center, Heading, IconButton, useDisclosure } from "@chakra-ui/react";
+import { Box, Center, Heading } from "@chakra-ui/react";
 import CardsStat from "../components/dashboard/cardsStat";
 import MainCard from "../components/dashboard/mainCard/mainCard";
