+++ frontend/utils/definitions/useQueryDefinition.ts
@@ -1,6 +1,6 @@
-// // for convention all methods created in this file will have at the start of the name the phrase "use" and on the end "Query"
+// // for convention, all methods created in this file will have at the start of the name the prefix "use" and on the end "Query"
 // // that's because all query have to use a query call, so they have to use the query method in queryDefinition.ts
-// // here you have to insert only the request but the call are to do on the component
+// // here you must insert only the request and the call needs to be done on the component
 
 import request from "graphql-request";
 import { queryUrl } from "../queryUrl";
@@ -32,6 +32,10 @@ import {
   TransactionsSearchQueryValueDefinition,
 } from "./typeDefinition";
 
+const headers = {
+  Authorization: `Bearer ${sessionStorage?.getItem("token")}`,
+};
+
 export const useCheckEmailExistQuery = async (emailFormValue: string) => {
   const res = request<UserExistByEmailQuery>(queryUrl, UserExistByEmailDocument, {
     email: emailFormValue,
@@ -62,10 +66,15 @@ export const useSignupQuery = async (props: SignupQueryValueDefinition) => {
   return res;
 };
 
-export const useUserByEmailQuery = async (props: GetUserByEmailQueryValueDefinition) => {
-  const res = await request<UserByEmailQuery>(queryUrl, UserByEmailDocument, {
-    email: props.email,
-  });
+export const UseUserByEmailQuery = async (props: GetUserByEmailQueryValueDefinition) => {
+  const res = await request<UserByEmailQuery>(
+    queryUrl,
+    UserByEmailDocument,
+    {
+      email: props.email,
+    },
+    headers
+  );
   props.setName(res.userByEmail.name);
   props.setSurname(res.userByEmail.surname);
   props.setEmail(res.userByEmail.email);
@@ -104,62 +113,82 @@ export const useUserByEmailQuery = async (props: GetUserByEmailQueryValueDefinit
 export const useAddOrUpdateTransactionQuery = async (props: TransactionQueryValueDefinition) => {
   if (!props.transaction) throw new Error("Missing transaction");
 
-  const res = await request<AddOrUpdateTransactionMutation>(queryUrl, AddOrUpdateTransactionDocument, {
-    transactionInput: {
-      userEmail: props.email,
-      accountId: props.accountId,
-      transaction: {
-        id: props.transactionId,
-        amount: parseFloat(props.transaction.amount.toString()),
-        currency: props.transaction.currency,
-        transactionType: props.transaction.operationType,
-        description: props.transaction.description,
-        dateTime: props.transaction.date,
-        category: {
-          id: props.transaction.category.id,
-          name: props.transaction.category.name,
-          categoryType: props.transaction.operationType,
+  const res = await request<AddOrUpdateTransactionMutation>(
+    queryUrl,
+    AddOrUpdateTransactionDocument,
+    {
+      transactionInput: {
+        userEmail: props.email,
+        accountId: props.accountId,
+        transaction: {
+          id: props.transactionId,
+          amount: parseFloat(props.transaction.amount.toString()),
+          currency: props.transaction.currency,
+          transactionType: props.transaction.operationType,
+          description: props.transaction.description,
+          dateTime: props.transaction.date,
+          category: {
+            id: props.transaction.category.id,
+            name: props.transaction.category.name,
+            categoryType: props.transaction.operationType,
+          },
         },
       },
     },
-  });
+    headers
+  );
   return res;
 };
 
 export const useDeleteTransactionQuery = async (props: TransactionQueryValueDefinition) => {
-  const res = await request<DeleteTransactionMutation>(queryUrl, DeleteTransactionDocument, {
-    transaction: {
-      userEmail: props.email,
-      accountId: props.accountId,
-      transactionId: props.transactionId,
+  const res = await request<DeleteTransactionMutation>(
+    queryUrl,
+    DeleteTransactionDocument,
+    {
+      transaction: {
+        userEmail: props.email,
+        accountId: props.accountId,
+        transactionId: props.transactionId,
+      },
     },
-  });
+    headers
+  );
   return res;
 };
 
 export const useDeleteTransactionListQuery = async (props: TransactionListQueryValueDefinition) => {
-  const res = await request<DeleteTransactionListMutation>(queryUrl, DeleteTransactionListDocument, {
-    transactions: {
-      userEmail: props.email,
-      accountId: props.accountId,
-      transactionIds: props.transactionIds,
+  const res = await request<DeleteTransactionListMutation>(
+    queryUrl,
+    DeleteTransactionListDocument,
+    {
+      transactions: {
+        userEmail: props.email,
+        accountId: props.accountId,
+        transactionIds: props.transactionIds,
+      },
     },
-  });
+    headers
+  );
   return res;
 };
 
 export const useTransactionsFilteredQuery = async (props: TransactionsSearchQueryValueDefinition) => {
-  const res = await request<UserTransactionsFilteredQuery>(queryUrl, UserTransactionsFilteredDocument, {
-    filters: {
-      userEmail: props.email,
-      accountId: props.accountId,
-      transactionFilters: {
-        dateStart: props.dateStart,
-        dateEnd: props.dateEnd,
-        categoriesIds: props.categoriesIds,
-        currencies: props.currencies,
+  const res = await request<UserTransactionsFilteredQuery>(
+    queryUrl,
+    UserTransactionsFilteredDocument,
+    {
+      filters: {
+        userEmail: props.email,
+        accountId: props.accountId,
+        transactionFilters: {
+          dateStart: props.dateStart,
+          dateEnd: props.dateEnd,
+          categoriesIds: props.categoriesIds,
+          currencies: props.currencies,
+        },
       },
     },
-  });
+    headers
+  );
   return res;
 };
